Move Sidebar to own file
Move the sidebar view to its own file and add a very basic test.
