+++ web/static/js/__tests__/Sidebar-test.jsx
@@ -1,0 +1,18 @@
+/** @jsx React.DOM */
+
+"use strict"
+
+var React = require('react');
+
+var Sidebar = require('../Sidebar.jsx');
+
+describe('Sidebar', function() {
+  it('renders without throwing', function() {
+    var query = 'cat:code-display tag:css';
+    var container = document.createElement('div');
+    React.renderComponent(
+      <Sidebar query={query} />,
+      container
+    );
+  });
+});
